+++ demo/stores/app.js
@@ -12,6 +12,11 @@ let data = ImmutableHelper.parseJSON({
   button: {
     text: "Action"
   },
+  checkbox: {
+    label: "Checkbox",
+    labelHelp: "Example label help.",
+    fieldHelp: "Example field help."
+  },
   confirm: {
     title: 'Are you sure?'
   },
@@ -22,10 +27,25 @@ let data = ImmutableHelper.parseJSON({
     title: 'Example Content',
     body: 'This is some example content\nfor the Content component.'
   },
+  date: {
+    label: "Date",
+    labelHelp: "Example label help.",
+    fieldHelp: "Example field help."
+  },
+  decimal: {
+    label: "Decimal",
+    labelHelp: "Example label help.",
+    fieldHelp: "Example field help."
+  },
   dialog: {
     open: false,
     title: 'Hello World!',
     disableBackground: true
+  },
+  dropdown: {
+    label: "Dropdown",
+    labelHelp: "Example label help.",
+    fieldHelp: "Example field help."
   },
   dialog_full_screen: {
     open: false,
@@ -36,6 +56,9 @@ let data = ImmutableHelper.parseJSON({
     as: "warning",
     open: true,
     text: "Sample flash notification."
+  },
+  help: {
+    message: "This is an example of a help tooltip."
   },
   icon: {
     type: 'info'
@@ -65,11 +88,9 @@ let data = ImmutableHelper.parseJSON({
     as: "primary"
   },
   radio_button: {
-    columnData: [
-      { name: 'access', value: 'full', label: 'Full Access'},
-      { name: 'access', value: 'limited', label: 'Limited Access'},
-      { name: 'access', value: 'no-access', label: 'No Access'}
-    ]
+    label: "Radio Button",
+    labelHelp: "Example label help.",
+    fieldHelp: "Example field help."
   },
   rainbow: {
     data: [{
@@ -123,11 +144,19 @@ let data = ImmutableHelper.parseJSON({
   tabs: {
     tabData: [{}, {}]
   },
+  textbox: {
+    label: "Textbox",
+    labelHelp: "Example label help.",
+    fieldHelp: "Example field help."
+  },
   toast: {
     dismissable: true,
     as: "warning",
     open: true,
     text: "Sample toast notification."
+  },
+  tooltip: {
+    message: 'Some Helpful Content'
   }
 });